challenges/web/ScreenshotMe/challenge/requirements.txt
Flask==2.0.2
selenium==4.1.0
Pillow==9.0.0
webdriver_manager==3.4.2
